requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
streamlit==1.31.0
langchain==0.1.6
langchain-openai==0.0.5
openai==1.12.0
python-dotenv==1.0.0
Pillow==10.2.0
SpeechRecognition==3.10.1
pydub==0.25.1
gTTS==2.5.1
requests==2.31.0
aiofiles==23.2.1
python-jose==3.3.0
pydantic==2.6.0
pydantic-settings==2.1.0